==> scripts/07_extra_tinytex-troubleshooting.R <==
#### Preamble ####
# Purpose: Troubleshooting Knitting in other devices
# Author: Jeremy Chu
# Contact: jeremychuj@gmail.com
# Date: 30 January 2021
# Pre-requisites: None. This is just for people who want to knit the .Rmd file but is running into TinyTex issues
# TODOs: -

## Troubleshooting link ##
# https://yihui.org/tinytex/r/#debugging

# Follow the instructions above to go through the tinytex troubleshooting
# I find updating R and LaTeX packages, and then reinstalling tinytex worked for me.
# If you get file corurpted errors, please restart the R session.
